CRS-1101 Remove validation call from the get-calc-breakdown endpoint, this is unnecessary and causes problems for the view journey
