+++ src/test/kotlin/uk/gov/justice/digital/hmpps/calculatereleasedatesapi/service/BookingServiceTest.kt
@@ -6,11 +6,17 @@ import org.junit.jupiter.api.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.reset
 import org.mockito.kotlin.whenever
+import org.springframework.http.MediaType
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.AdjustmentType.REMAND
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.AdjustmentType.UNLAWFULLY_AT_LARGE
+import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.enumerations.CalculationStatus.PRELIMINARY
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Adjustment
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Adjustments
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Booking
+import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculatedReleaseDates
+import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationBreakdown
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationSentenceUserInput
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.CalculationUserInputs
 import uk.gov.justice.digital.hmpps.calculatereleasedatesapi.model.Duration
@@ -391,6 +397,34 @@ class BookingServiceTest {
       )
     )
   }
+
+  @Test
+  fun `Test GET of calculation breakdown by calculationRequestId`() {
+    val sentences: List<SentenceAndOffences> = emptyList()
+    val adjustments = BookingAndSentenceAdjustments(emptyList(), emptyList())
+    val calculationRequestId = 9995L
+    val offender = Offender(prisonerId, LocalDate.of(1980, 1, 1))
+    val booking = Booking(offender, mutableListOf(), Adjustments(), null, bookingId)
+    val breakdown = CalculationBreakdown(listOf(), null)
+    val calculation = CalculatedReleaseDates(
+      calculationRequestId = 9991L, dates = mapOf(), calculationStatus = PRELIMINARY,
+      bookingId = bookingId, prisonerId = prisonerId
+    )
+    val userInput = CalculationUserInputs(listOf(CalculationSentenceUserInput(1, "ABC", UserInputType.ORIGINAL, true)))
+
+    whenever(calculationTransactionalService.findCalculationResults(calculationRequestId)).thenReturn(calculation)
+    whenever(calculationTransactionalService.findSentenceAndOffencesFromCalculation(calculationRequestId)).thenReturn(sentences)
+    whenever(calculationTransactionalService.findPrisonerDetailsFromCalculation(calculationRequestId)).thenReturn(prisonerDetails)
+    whenever(calculationTransactionalService.findOffenderFinePaymentsFromCalculation(calculationRequestId)).thenReturn(offenderFineFinePayment) // TODO
+    whenever(calculationTransactionalService.findBookingAndSentenceAdjustmentsFromCalculation(calculationRequestId)).thenReturn(adjustments)
+    whenever(calculationTransactionalService.findUserInput(calculationRequestId)).thenReturn(userInput)
+    whenever(calculationTransactionalService.calculateWithBreakdown(booking, calculation)).thenReturn(breakdown)
+
+    val result = bookingService.getCalculationBreakdown(calculationRequestId)
+
+    assertThat(result).isEqualTo(breakdown)
+  }
+
   private companion object {
     val FIVE_YEAR_DURATION = Duration(mutableMapOf(DAYS to 0L, WEEKS to 0L, MONTHS to 0L, YEARS to 5L))
     val FIRST_JAN_2015: LocalDate = LocalDate.of(2015, 1, 1)
